refactor:apiStudentSkills
