+++ src/modules/student/features/useAddStudentSkills.js
@@ -1,6 +1,7 @@
 import { useMutation, useQueryClient } from "@tanstack/react-query";
-import { addStudentSkills } from "../api/addStudentSkills";
+
 import toast from "react-hot-toast";
+import { addStudentSkills } from "../api/apiStudentSkills";
 
 export function useAddStudentSkills() {
   const queryClient = useQueryClient();
